RaspberryPi_Code/max31855/requirements.txt
pymodbus==3.7.4
RPi.GPIO==0.7.1
spidev==3.6